feat: 정수뿐만 아니라 실수도 연산 할 수 있도록 변경. long -> double
- 연산 결과가 정수로만 출력되는 문제 해결.
- 실수도 연산할 수 있도록 변경.

+++ src/main/java/calculator/Application.java
@@ -12,7 +12,7 @@ public class Application {
     public static void main(String[] args) throws IOException {
         while (true) {
             String[] splitUserInput;
-            long result;
+            double result;
 
             try {
                 System.out.print("식을 입력해 주세요(그만하려면 q): ");
